Fix github workflow (#39)
* Fix github workflow

Failing due to changes in Ansible requirements files.

* Fix github workflow

Try again, only support newer versions of Python.

* Fix github workflow

Comment out linting of playbooks due to bug in ansible-lint. Hope to be
able to restore this in future.

* Fix github workflow

Seems as if python version 3.10 is interpreted as 3.1. Remove it from
list.

* Fix github workflow

Another instance of a now remvoed requirements file.

* Fix github workflow

Fix check_required_together (from the porting guide).