Rmb openflow (#41)
* Openflow: Refactor as resource module

Refactor OpenFlow support as a resource module. Add support for
crypto parameters.

* Fix formatting

Github workflow picked up some pep8 errors.

* Fix minor errors

Parsing model and serial number when model name has an embedded space, as
for the "AT-SBx908 GEN2". Generating facts when there is no configuration
for the module. Generating facts when the module is not supported on the
target, as is the case for BGP on the x230.

* Fix minor formatting error

* Fix openflow state "replaced"

An unexpected error was caused by not generating the correct commands
for a state of "replaced". Add some tests to cover this case.

* Fix deletion of missing OpenFlow controller

No need to generate an error, we will just silently accept this,
which is the general approach taken.