Fixes to allow Galaxy workflow to pass
A number of failures when submitting the collection to Galaxy need to
be addressed, including a number of flake8 issues and issues with
documentation strings.

Also added a new file meta/runtime.yml which seems to be required now,
and with a "requires_ansible" tag in it.

Also add options to galaxy.yml to exclude some files from the collection
build.